Test for Device Object Management.

+++ mbed-client/m2mconstants.h
@@ -143,7 +143,7 @@ const uint8_t LENGTH8 = 0x08;
 const uint8_t LENGTH16 = 0x10;
 const uint8_t LENGTH24 = 0x18;
 
-const int COAP_CONTENT_OMA_TLV_TYPE = 99;
-const int COAP_CONTENT_OMA_JSON_TYPE = 100;
+const uint8_t COAP_CONTENT_OMA_TLV_TYPE = 99;
+const uint8_t COAP_CONTENT_OMA_JSON_TYPE = 100;
 
 #endif // M2MCONSTANTS_H
